+++ src/test/java/org/folio/entitlement/support/TestConstants.java
@@ -49,6 +49,7 @@ public class TestConstants {
   public static final UUID FLOW_ID = UUID.randomUUID();
   public static final UUID APPLICATION_FLOW_ID = UUID.randomUUID();
   public static final UUID STAGE_UUID = UUID.randomUUID();
+  public static final UUID OWNER_INSTANCE_ID = UUID.randomUUID();
   public static final String FLOW_STAGE_ID = FLOW_ID + "/ApplicationsFlow/Level-0/" + APPLICATION_FLOW_ID;
 
   public static final HttpClient HTTP_CLIENT_DUMMY_SSL = httpClientWithDummySslContext();